Corner platform Pregen
Fixed up the corner prefab as well as allowed the platform gates to
spawn the corner platform.

Also made the 'wallrunning' platform prefab, but not too happy with it,
and the actual running along the wall part doesn't work. Couldn't figure
out a way to make the player not fall from gravity while holding 'jump'
without also making him shoot into the sky.