Path Generator Fix (Panic State), Some UI Changes
The path generator should no longer stop when it has met a path that has
not dropped for some reason - it now has a panic state, which removes
obstructing tiles when the player is atop the last tile that was
spawned. We should hopefully be free of those game-breaking path
generation bugs!